refactor(auth): move api service to services folder and fix authtype reference naming
